wip: move more entities
