+++ custom_components/venta/select.py
@@ -2,23 +2,14 @@
 
 from __future__ import annotations
 
-from collections.abc import Callable
-from dataclasses import dataclass
-
-from homeassistant.components.select import (
-    SelectEntity,
-    SelectEntityDescription,
-)
 from homeassistant.config_entries import ConfigEntry
-from homeassistant.const import (
-    EntityCategory,
-)
+from homeassistant.const import EntityCategory
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
-from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
 from .const import ATTR_LED_STRIP_MODE, DOMAIN
-from .venta import VentaData, VentaDataUpdateCoordinator
+from .venta import VentaDataUpdateCoordinator
+from .venta_entity import VentaSelect, VentaSelectEntityDescription
 
 LED_STRIP_MODES = {
     0: "internal",
@@ -28,22 +19,6 @@ LED_STRIP_MODES = {
 }
 LED_STRIP_MODES_KEYS = list(LED_STRIP_MODES.keys())
 LED_STRIP_MODES_VALUES = list(LED_STRIP_MODES.values())
-
-
-@dataclass
-class VentaSelectRequiredKeysMixin:
-    """Mixin for required keys."""
-
-    exists_func: Callable[[VentaDataUpdateCoordinator], bool]
-    value_func: Callable[[VentaData], str | None]
-    action_func: Callable[[str], dict | None]
-
-
-@dataclass
-class VentaSelectEntityDescription(
-    SelectEntityDescription, VentaSelectRequiredKeysMixin
-):
-    """Describes Venta select entity."""
 
 
 SENSOR_TYPES: list[VentaSelectEntityDescription] = (
@@ -79,34 +54,3 @@ async def async_setup_entry(
         if description.exists_func(coordinator)
     ]
     async_add_entities(entities)
-
-
-class VentaSelect(CoordinatorEntity[VentaDataUpdateCoordinator], SelectEntity):
-    """Representation of a select."""
-
-    _attr_has_entity_name = True
-    entity_description: VentaSelectEntityDescription
-
-    def __init__(
-        self,
-        coordinator: VentaDataUpdateCoordinator,
-        description: VentaSelectEntityDescription,
-    ) -> None:
-        """Initialize the sensor."""
-        super().__init__(coordinator)
-        self.entity_description = description
-        self._attr_device_info = coordinator.device_info
-        self._attr_unique_id = f"{coordinator.api.device.mac}-{description.key}"
-        self._attr_options = description.options
-
-    @property
-    def current_option(self) -> str | None:
-        """Return the selected entity option to represent the entity state."""
-        return self.entity_description.value_func(self.coordinator.data)
-
-    async def async_select_option(self, option: str) -> None:
-        """Change the selected option."""
-        await self.coordinator.api.device.status(
-            self.entity_description.action_func(option)
-        )
-        await self.coordinator.async_request_refresh()
